testing/web-platform/mozilla/tests/infrastructure/specialPowers/specialpowers.html

<title>Check specialPowers is available in gecko-only tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
test(() => assert_equals(SpecialPowers.sanityCheck(), "foo"))
</script>
